requirements-dev.txt
Pillow==10.3.0
pytest==8.2.1
